Added reap_set_state function
btw, this fixes a bug with log-expl-time, because timers were not
started in reap_init_explore. Now reap_set_state is responsible for
state transition and associated actions. It is not allowed anymore to
make a state transition manually, which ensures that we do not forget
mandatory actions upon state transition (log-expl-time is one such
example).