BUG FIX:Fixed error management in log_expl_time
bug:If the /etc/shim6 directory did not exist, an error message was
given, but reapd failed to switch to the working path.
fix:We now try to create the /etc/shim6 when the open fails with ENOENT,
then try again to open the log file. If this fails again, then we report
that we cannot log the expl time, but shim6 continues working correctly.