BUG FIX: (tcp_hint) Bad exploration order
the cur_path_index field of rctx was not reset to 0 after randomization.
While this is not a problem in case of pure randomization of the paths,
it becomes a problem there is some intelligent ordering of the pairs.
In the tcp_hint branch, we separate the paths in two sets. The first set
holds the paths that are completely distinct from the current one (in
terms of the address pair). The second set holds the paths that have at
least one address in common with the current path. The idea is to first
explore maximally distinct paths. In that case, we need to reset the
cur_path_index to 0, so as to ensure that the first set is actually
explored first.