BUG FIX: UR not accepted sometimes+sent probe list not cleared upon update
The UR was sometimes not accepted because CGA PDS verif was against
source locator, instead of against peer ULID. So, if the peer sent the
UR with another locator than the ULID, the UR was incorrectly dropped at
reception.

Other bad uses of function shim6_is_remote_cga are also fixed in this
patch.

The other bug fix contained in this patch is about probe list. The list
must be cleared when completing an update.