Added ability to handle 0-sized local locator sets.
0-sized locator sets can really happen when LinShim6 is used with MIPv6.
A typical case is when the only available address is moved from the network interface
to the tunnel interface. MIPv6 first removes the address, then immediately adds the same
address to the tunnel.

Another case is when a router entry expires in MIPv6. This can happen even if normal
RAs are sent without problem, because by default MIPv6 sets its router expiry time
to the RA interval, in order to faster discover movements. However, if the Router
Advertisement entity makes a random variation around the advertised time, which is usual,
it will happen sometimes that MIPv6 will decide that its router entry has expired, and
that it is time to check if we are still on the same link. This triggers a DAD on the
current address (the only one again). The technique used by the MIPv6 daemon to trigger a DAD
over its current address is to delete the address, then add it again. That way the kernel
does the DAD itself, but LinShim6 sees a transient situation with no local locator at all.

This patch solves these two issues.