Added support for dynamically changing interfaces
When an interface is removed , its addresses are also removed.
Thus the daemon is notified, with the address and ifidx concerned.
But shim6d finds that the ifix is no longer valid, since the interface
has been removed. Previously this caused shim6d to abord with assertion
failed. Now we take this into account, and search through all lists of
addresses, in order to actually delete it. Note that we make the
assumption that a given address is not configured in Shim6d on several
interfaces in the same time.